完成认证权限

+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -33,7 +34,9 @@ class PostController extends Controller
         ]);
 
         // 逻辑
-        $post = Post::create(request(['title', 'content']));
+        $user_id = Auth::id();
+        $params = array_merge(request(['title', 'content']), compact('user_id'));
+        $post = Post::create($params);
 
         // 渲染
         return redirect('/posts');
@@ -53,6 +56,7 @@ class PostController extends Controller
             'content' => 'required|string|min:10',
         ]);
 
+        $this->authorize('update', $post);
         // 逻辑
         $post->title = request('title');
         $post->content = request('content');
@@ -63,7 +67,8 @@ class PostController extends Controller
 
     public function delete(Post $post)
     {
-        // TODO: 用户权限认证
+        $this->authorize('delete', $post);
+
         $post->delete();
         return redirect("/posts");
     }
